安装时产生mkcp+随机3种伪装类型type的配置文件
