+++ changestream.py
@@ -4,26 +4,6 @@ import readjson
 import writejson
 import v2rayutil
 
-def writeStreamJson(newstreamnetwork):
-	if(newstreamnetwork=="1"):
-	    writejson.WriteStreamNetwork("tcp","none")
-	elif(newstreamnetwork=="2"):
-	    print("请输入你想要为伪装的域名（不不不需要http）：")
-	    host=raw_input()
-	    writejson.WriteStreamNetwork("tcp",str(host))
-	elif(newstreamnetwork=="3"):
-	    print("请输入你的服务器绑定域名（不不不需要http）：")
-	    host=raw_input()
-	    writejson.WriteStreamNetwork("ws",str(host))
-	elif(newstreamnetwork=="4"):
-	    writejson.WriteStreamNetwork("mkcp","none")
-	elif(newstreamnetwork=="5"):
-	    writejson.WriteStreamNetwork("mkcp","kcp srtp")
-	elif(newstreamnetwork=="6"):
-	    writejson.WriteStreamNetwork("mkcp","kcp utp")
-	elif(newstreamnetwork=="7"):
-	    writejson.WriteStreamNetwork("mkcp","kcp wechat-video")
-
 #读取配置文件信息
 mystreamnetwork=str(readjson.ConfStreamNetwork)
 if readjson.ConfStreamNetwork=="kcp" :
@@ -60,7 +40,7 @@ if ( not v2rayutil.is_number(newstreamnetwork)):
     exit
 else:
     if not (newstreamnetwork > 0 and newstreamnetwork<8):
-    	writeStreamJson(newstreamnetwork)
+    	v2rayutil.writeStreamJson(newstreamnetwork)
     else:
         print("请输入有效数字！")
         exit
