+++ v2rayutil.py
@@ -1,5 +1,7 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
+import writejson
+import random
 
 #判断是否为数字的函数
 def is_number(s):
@@ -17,3 +19,27 @@ def is_number(s):
         pass
  
     return False
+
+def writeStreamJson(newstreamnetwork):
+    if(newstreamnetwork=="1"):
+        writejson.WriteStreamNetwork("tcp","none")
+    elif(newstreamnetwork=="2"):
+        print("请输入你想要为伪装的域名（不不不需要http）：")
+        host=raw_input()
+        writejson.WriteStreamNetwork("tcp",str(host))
+    elif(newstreamnetwork=="3"):
+        print("请输入你的服务器绑定域名（不不不需要http）：")
+        host=raw_input()
+        writejson.WriteStreamNetwork("ws",str(host))
+    elif(newstreamnetwork=="4"):
+        writejson.WriteStreamNetwork("mkcp","none")
+    elif(newstreamnetwork=="5"):
+        writejson.WriteStreamNetwork("mkcp","kcp srtp")
+    elif(newstreamnetwork=="6"):
+        writejson.WriteStreamNetwork("mkcp","kcp utp")
+    elif(newstreamnetwork=="7"):
+        writejson.WriteStreamNetwork("mkcp","kcp wechat-video")
+
+
+def randomStream():
+    writeStreamJson(str(random.randint(5,7)))
